✨ 统一事件模型
